Fix multiline string in help_text
When migrating my mediathread db, I was getting the message that told me to
run "makemigrations". When I did that, it created a migration for
Asset.metadata_blob with slightly different spacing in the help text.

Using a different method for making the multi-line string in this commit
prevents the migration prompt from appearing.