faculty_only mixin implies logged in user
see: https://sentry.ccnmtl.columbia.edu/sentry-internal/mediathread/group/764/

if an anon hits a view protected only with faculty_only, we get a server error instead of a 403.

this fixes that issue.

arguably, views with @faculty_only probably ought to also have the @login_required
decorator so the user gets sent to the login page rather than a 403, but this ensures
that even if that isn't in place, we get a 403 instead of a 500