Fix missing clipform selection buttons - PMT #102158
After lots of debugging, I found out how the clipform widget
is getting inserted into the dom. The problem here was that the
elements it was trying to append to weren't available yet with
my new Mustache refactor. I've updated the necessary parts of code
so sherd is able to insert the form into the quick edit modal.