Integrate relevant parts of sherdjs into mediathread
This adds the relevant parts of sherdjs (everything except
the chrome extension stuff and the bookmarklet) into
the mediathread repository to make development less awkward.

I haven't touched the sherdjs tarball dependency - for the time
being we can still use that for accessing the bookmarklet, but
that can go away soon too.. this is just the first step.

Eventually, I'm imagining the bookmarklet code forked into its own
repository.. and the SherdJS repository going away, until we find
some useful abstraction layer to work in. I don't think we're
losing anything by putting this code into mediathread: We
can still copy this code out into other applications. If that
happens, then we can think about generalizing the implementation.

I'm open to alternative methods of course.. I just want to get the
process moving because this will make it simpler to add fixes to the
asset viewers in mediathread.