Add link to install chrome extension - PMT #103376
I followed these instructions to add a button to mediathread
to install the extension:
https://developer.chrome.com/webstore/inline_installation

Note that because this only works on "Verified domains", you
need to be on a *.ccnmtl.columbia.edu domain to test this.

Because http://ccnmtl.columbia.edu/ happened to be registered
under google webmaster tools as the ccnmtl@gmail.com user,
I was able to connect it up to the chrome extension in the developer
dashboard.