Update source images for external collections
This updates a handful of source images for the
ExternalCollections that are pointing to mediathread's
local images. I'm also using larger sizes so they can be
scaled to the appropriate resolution on retina displays.

I've updated the CSS to not change the aspect ratio for these
logos. I think it looks worse to squish the image than to
just have a tiny logo if it's not square (like many of the
logos getting linked to at external domains).

There are many ExternalCollections that aren't pointing to
local image files -- we could update those with a data
migration in the future. We can talk about that.