Fix Assignment feature when testing in Firefox
When running this test in Firefox, it throws a missing selector error when searching for `td.panel-container.open.composition`. When it's in this state, there is an open Assignment panel, not Composition. I ran into this when trying to debug a different error on my bootstrap buttons branch.

To reproduce this problem, change the test browser in `settings_test.py` to Firefox, and run `./manage.py harvest --settings=mediathread.settings_test mediathread/projects/features/assignment.feature --failfast --pdb`.