Vimeo cleanup
Clean up some vimeo initialization stuff - pass 'autoplay'
param to vimeo embed.

We talked about how the videos on Composition view aren't supposed
to autoplay. The parameter that controls this is here:
https://github.com/ccnmtl/mediathread/blob/master/media/js/lib/sherdjs/src/configs/djangosherd.js#L256

Setting that to 'false' makes these videos not autoplay when clicked on,
however, that also makes selections on this view not get initialized
in the right place, for youtube and vimeo videos, so I'm leaving
this set to true for now.