Standalone Video Play Button
Meth videos are normally displayed with a clipform that contains a "play" button along
with a set of selection tools. The play button seeks to the correct selection location
and plays the video. In cases where the clipform is not displayed, for example: the LTI
embed selection form, setup a prominent button to allow users to quickly play the selection.