Course activation preliminary - PMT #105457
CourseActivateView basics

Renamed `LoggedInMixin` to `LoggedInCourseMixin`. `LoggedInMixin` is now
a simple check to see if the user is logged in - doesn't rely on course
stuff.

The ActivatableAffil model stores affiliations that the user may want to
activate into a Mediathread course. There are checks in auth.py to see
if the user is a faculty-member of this affiliation, if the date makes
sense, etc. I'll fine-tune these conditions as we go.

There's a simple course-activation form that's not functional yet.

I've introduced MethCourseManagerMiddleware that forwards the user to
this new 'homepage' instead of the old one, if the waffle flag is
active for them.