Update WGBH vietnam thumbnail URL
This url was pointing to http://ccnmtl.columbia.edu/ when we can
just host this image in mediathread. It looks like this doesn't
actually fix the mixed-content warning I'm seeing in Chrome, but
in any case this should be fine host locally.

A related question: How and when is supported_archives.py used?
grepping the codebase for supported_archives didn't bring anything
up.