revive the docker/docker-compose setup
I had previously commented out the test running step of building the
docker container because Mediathread's tests (for some undetermined
reason) would consistently fail on my desktop, so it was the only way I
could get it to build. That issue has resolved itself though, so I put
that back in as a nice sanity check within the container.

I've also updated the documentation a bit with more details on how to
use docker-compose to run a development environment.

Lastly, I tend to run into issues with running Mediathread locally
because of the Pickle session serializer, which always breaks because my
browser generally already has a (JSON) session cookie for my local
setup. Clearing out the cookies is a pain, so I set the serializer in
the compose settings to JSON serializer to make it convenient to do
development against it. Since no one else is using the compose setup yet
that I know of, it shouldn't break anyone else's setup.