Course activation member fix
This should address the problem where users weren't being added to the
course that they activate. I was adding them to the course's
`faculty_group`, but not the course's `group`.

I changed the test course name form "My Course" to "English for Cats" to
avoid any confusion in the assertions because the text "My Courses" is
also present on this page.

In order to properly test this, the Affil needs to be generated as
`.fc.` instead of `.st.` - that's why I have that factory change in
here.