Only allow sequence submission when valid - PMT #109151
This prevents the Submit button from being active when the sequence is
in a non-submittable state.

We'll also be setting the submittable state from React as the project
loads, during initialization. This is for the case where the response
is in a submittable state, but the user has just navigated to it without
doing any editing or saving.