Fix sequence serializer rounding exception - PMT #109487
An end_time or start_time of `None` should cause an error in the API,
not an exception in the code.

I've also updated the serializer to do the rounding with the Decimal
type since that is how these timecodes are ultimately represented.
Using `round()` converts to a float, adding another step in the process
and more room for error.