Filter out Vimeo assets from CollectionWidget - PMT #109584
I made the primary_type filter accept an array as well as a single
value. It looks like we need to filter out 'image_fpxid' (artstor) as well
as the 'vimeo' primary type.

This is my first foray into the asset filtering code and the
RestrictedFiltersMixin so please review and edit or make suggestions as
needed.