Set up sphinx documentation
I'm setting up Sphinx documentation for the Mediathread, and I'm hoping
that having the documentation closer to the code will make it easier for
us to keep up to date. For example, I was reading through the "Supported
Media" page and realized it should have been updated a while ago, when I
decided to set this up.