Enable sequence composition for all users
Note: sequence compositions don't have any requirements in order to be
save, so a user can publish an empty sequence to the world.

When viewing a public sequence composition that doesn't have a spine
video, the "Add a primary video" box appears, and the viewer can pick
the spine video! Honestly, I think this is a fine easter egg, because
no one is going to publish an empty sequence, share it with someone and
really need them to see an empty box.