LTI: automatic course provisioning
Courses can be automatically created or connected in one step for our various
ecosystems. The view keys off a custom lti parameter to structure expected
student and faculty affiliation groups. In the event that the custom parameter
is missing or in an unexpected format, the code defaults to the unique
custom_course_context parameter.