:arrow_up: OpenLayers 2.13.1
So, it's not clear what version of OpenLayers mediathread is currently
on (`VERSION_NUMBER` is set to the value `$Revision$`), but based on the
git history in sherdjs, it appears it's somewhere around 2.x. I've
updated this to the most recent version available for 2.x: 2.13.1.

OpenLayers 2.x isn't compatible with jQuery 3.x, likely because of
something trivial. So we'll have to get to at least openlayers 3 to use
jQuery 3.