Update browser extension requirements
The extension actually works with any browser based on Chromium (I use
it in Brave):
https://en.wikipedia.org/wiki/Chromium_(web_browser)#Browsers_based_on_Chromium

So, I don't want users of other browsers to feel left out here. I think
"a browser in the Chrome family" is succint enough, and still makes it
obvious for less technical users that they just need to use Chrome.

Changing this line:

    You must be using Chrome to install this extension.

to:

    You must be using a browser in the Chrome family to install this
    extension.