Add panel fetch error if panel load ajax call fails
I just worked through a bug where it's possible for the course to not be
set when loading the project workspace. You can work around this bug by
going to course select, selecting your course, and going back to the
project view.

I haven't resolved the bug yet, but adding an error here will help with
debugging, as it was pretty confusing to see what was going on with no
errors and a blank page.