Discussions form: exclude a form that has the "response" class
When testing this between development and production, I noticed this
particular discussion on production is initiated with two slightly
different "threaded_comments_form" forms. One of them has the "response"
class added - and this is getting selected as the DiscussionPanel's
self.form in the front-end code, breaking the response.

This should fix the issue, but I'm curious how this is happening in the
first place.