Project.faculty_compositions: allow faculty_group to be None
This code should not fail when this value isn't set. This was uncovered
in development testing with a very minimal course and wouldn't
necessarily happen in production. I've also added a unit test for this.

Also, this function is essentially a static method that doesn't use
`self`, so I've update the function signature.