Move base_new template to base.html - closes #5571
base.html is an old, deprecated template which can be removed.

In this pull request, I'm leaving the base_new.html file in place,
because a few places in mediathread_deploye_specific actually inherit
from this template. But once this is merged, I'll make a new pull
request to remove it.