Introduce django-lti - INFRA-2583
For LTI 1.3 compatibility.

An "LTI registration" object can be created in the lti_tool section of
Django admin. This enables the view to be linked to from Canvas.

Here are some of the configuration options which must be filled in when
creating an LTI registration within the Django admin:
* Client ID: The ID of the Developer Key created in Canvas
* Auth URL: `https://<canvas hostname>/api/lti/authorize_redirect`
* Access token URL: `https://<canvas hostname>/login/oauth2/auth`
* Keyset URL: `https://<canvas hostname>/api/lti/security/jwks`

I am not 100% certain on the accuracy of all of these values yet. I am
still investigating the integration steps. It seems many different
integration methods are possible.