Start AssetDetail refactor to function component
I want to update this giant component to a more modern React function
component, to allow for some react-player updates. This refactor will
take some time and a lot of steps - this is just the beginning and
separating out a minor function that I was able to, and creating a
wrapper function component.